build: use define_variable with pkgconfig paths
On NixOS all packages are installed into their own immutable prefix.
Because of this get_pkgconfig_variable will return a
path from within the package's prefix and we cannot write to it.
By using define_variable we can replace the respective directory
to be from the paths from meson. This should have no affect
on elementaryOS.